Testing ResourceInfo for single file
This is some late night testing to gather ResourceInfo for a single file.
In my case the squirrel demo file.

Finally found the reason I'm not getting the data i want/need. The ResourceInfo just returns data on folders, that need to be sorted out.
Next attempt will be to use the target to get the data by searching the parent for the target to get the resourceInfo that is needed.